# Document state storage