--- a/src/Dependencies/Collections/ImmutableSegmentedDictionary`2+Builder.cs
+++ b/src/Dependencies/Collections/ImmutableSegmentedDictionary`2+Builder.cs
@@ -181,6 +181,9 @@ namespace Microsoft.CodeAnalysis.Collections
                 return true;
             }
 
+            public void Compact()
+                => GetOrCreateMutableDictionary().Compact();
+
             public void RemoveRange(IEnumerable<TKey> keys)
             {
                 if (keys is null)