--- a/src/Tools/IdeCoreBenchmarks/ProjectOperationBenchmarks.cs
+++ b/src/Tools/IdeCoreBenchmarks/ProjectOperationBenchmarks.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Linq;
 using System.Text;
 using BenchmarkDotNet.Attributes;
 using Microsoft.CodeAnalysis;
@@ -13,6 +14,8 @@ namespace IdeCoreBenchmarks
 {
     public class ProjectOperationBenchmarks
     {
+        private static readonly SourceText s_newText = SourceText.From("text");
+
         [MemoryDiagnoser]
         public class IterateDocuments
         {
@@ -20,6 +23,7 @@ namespace IdeCoreBenchmarks
             private Project _emptyProject;
             private Project _hundredProject;
             private Project _thousandsProject;
+
 
             public IterateDocuments()
             {
@@ -96,6 +100,17 @@ namespace IdeCoreBenchmarks
                 }
 
                 return count;
+            }
+
+            [Benchmark(Description = "Solution.WithDocumentText")]
+            public void WithDocumentText()
+            {
+                var solution = Project.Solution;
+                var documentId = Project.DocumentIds.FirstOrDefault();
+                if (documentId != null)
+                {
+                    var _ = solution.WithDocumentText(documentId, s_newText);
+                }
             }
         }
     }