… src/Workspaces/Core/Portable/Workspace/Solution/ProjectChanges.cs
… src/Workspaces/Core/Portable/Workspace/Solution/ProjectState.cs
… src/Workspaces/Core/Portable/Workspace/Solution/ProjectState_Checksum.cs
… src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.CompilationAndGeneratorDriverTranslationAction_Actions.cs
… src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.CompilationTracker.State.cs
… src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.CompilationTracker.cs
… src/Workspaces/Core/Portable/Workspace/Solution/SolutionState.cs
… src/Workspaces/Core/Portable/Workspace/Solution/StateChecksums.cs
… src/Workspaces/Core/Portable/Workspace/Solution/TextDocumentStates.cs
… src/Workspaces/Core/Portable/Workspace/Workspace.cs
… src/Workspaces/CoreTest/SolutionTests/SolutionTests.cs
… src/Workspaces/Remote/ServiceHub/Host/RemoteWorkspace.SolutionCreator.cs
… src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Utilities/CompilerUtilities/ImmutableHashMapExtensions.cs