@CyrusNajmabadi 2021-01-12
@jasonmalinowski  any concern about:

> Instead, use the order the documents were added to the project.

We dont' have anything happening ambiently that would cause that order to change, right? i.e. when typing, it's not lke we remove projects/re-add to the end?