@CyrusNajmabadi 2021-01-12
> Add internal API Solution.GetDocumentAsync that retrieves a source document, regardless of whether it's source generated or not.

I don't think i like that.  Primarily because wouldn't that really be confusing when reviewing code?  In some places, `GetDocument` will return only non-generated files, but in others it can return generated files?  Honestly, i'd prefer we be consistent with the (wordy) GetAllRegularAndSourceGeneratedDocumentsAsync function as it's totally evident and helps bifurcate the API n clear lines.