@tmat 2021-01-12
@CyrusNajmabadi 
How about  `Solution.GetDocumentAsync(DocumentId id, bool includeSourceGenerated = false)` ?