@tmat 2021-01-22
@jasonmalinowski Re perf: Planning to update the PR to use segmented dictionary/array Sam is working on.