@tmat 2021-02-15
@CyrusNajmabadi @sharwell @jasonmalinowski PTAL - implemented optimizations on top of previous changes.